ENH:TST: Mass update to ILP64 compatible tests and partial z infra (#3)
[skip docs]

* ENH:TST: Convert all tests to ILP64 compatible and add z infra

* ENH:Fix all ILP64 related issues

* SCRIPT:Add the one time script used for posterity

* CI:Fix all double promotion warnings

+++ tests/chkxer.c
@@ -9,10 +9,10 @@
 #include <stdio.h>
 #include "verify.h"
 
-void chkxer(const char* srnamt, int infot, int* lerr, int* ok) {
+void chkxer(const char* srnamt, INT infot, INT* lerr, INT* ok) {
     if (!(*lerr)) {
-        fprintf(stderr, " *** Illegal value of parameter number %d not detected by %s ***\n",
-                infot, srnamt);
+        fprintf(stderr, " *** Illegal value of parameter number %lld not detected by %s ***\n",
+                (long long)infot, srnamt);
         *ok = 0;
     }
     *lerr = 0;
